requirements.txt
google-cloud-translate==2.0.1
pandas
spacy
scikit-learn
numpy
pillow
matplotlib
nltk
tqdm
torchinfo
torchtext==0.13.0
quickdraw
transformers
wandb